finally added full analysis
